# openapi.yaml
openapi: 3.0.3
info:
  title: Infinity Assistant API
  description: |
    Public-facing AI assistant API for InfinityAssistant.io
    
    Features:
    - Chat and conversation management
    - Knowledge base search and research
    - Memory and preferences
    - Analytics and tracking
    - Webhooks and integrations
    
    Authentication:
    - Email-based authentication
    - API key authentication (for developers)
  version: 1.0.0
  contact:
    name: Infinity Assistant Support
    url: https://infinityassistant.io/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://infinityassistant.io/api
    description: Production
  - url: http://localhost:3000/api
    description: Development
...
paths:
  /auth/email:
    get:
      tags: [Authentication]
      summary: Check authentication status
      responses:
        '200':
          description: Authentication status
          content:
            application/json:
              schema:
                type: object
                properties:
                  authenticated:
                    type: boolean
                  user:
                    type: object
    post:
      tags: [Authentication]
      summary: Sign in with email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Sign in successful
        '400':
          description: Invalid request
...
  /analytics/knowledge:
    get:
      tags: [Analytics]
      summary: Get knowledge analytics
      security:
        - ApiKeyAuth: []
        - EmailAuth: []
      parameters:
        - name: mode
          in: query
          schema:
            type: string
            enum: [all, professional, companion]
      responses:
        '200':
          description: Analytics data
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    EmailAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT